gateway/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.2
SQLAlchemy[asyncio]==2.0.31
asyncpg==0.29.0
redis==5.0.4
grpcio==1.64.0
grpcio-tools==1.64.0
protobuf==5.27.2
httpx==0.27.0
orjson==3.10.6
python-dotenv==1.0.1
loguru==0.7.2
